moved ai image generator into /work page
